Added a static method to TestRecipeBuilder to create TestRecipe from JSON text, rather than having a method in executor
