+++ readyapi4j-core/src/main/java/com/smartbear/readyapi4j/AssertionTypeResolver.java
@@ -1,4 +1,4 @@
-package com.smartbear.readyapi4j.testserver.execution;
+package com.smartbear.readyapi4j;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.DatabindContext;
@@ -21,23 +21,7 @@ import com.smartbear.readyapi.client.model.ValidHttpStatusCodesAssertion;
 import com.smartbear.readyapi.client.model.XPathContainsAssertion;
 import com.smartbear.readyapi.client.model.XQueryContainsAssertion;
 
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.GROOVY_SCRIPT;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.INVALID_HTTP_STATUS_CODES;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.JDBC_STATUS;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.JDBC_TIME_OUT;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.JSON_PATH_COUNT;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.JSON_PATH_MATCH;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.NOT_SOAP_FAULT;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.RESPONSE_SLA;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.SCHEMA_COMPLIANCE;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.SIMPLE_CONTAINS;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.SIMPLE_NOT_CONTAINS;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.SOAP_FAULT;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.VALID_HTTP_STATUS_CODES;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.X_PATH_MATCH;
-import static com.smartbear.readyapi4j.testserver.execution.AssertionNames.X_QUERY_MATCH;
-
-public class AssertionTypeResolver implements TypeIdResolver {
+class AssertionTypeResolver implements TypeIdResolver {
     private JavaType baseType;
 
     @Override
@@ -66,35 +50,35 @@ public class AssertionTypeResolver implements TypeIdResolver {
 
     private JavaType typeFromId(String type) {
         switch (type) {
-            case VALID_HTTP_STATUS_CODES:
+            case AssertionNames.VALID_HTTP_STATUS_CODES:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, ValidHttpStatusCodesAssertion.class);
-            case INVALID_HTTP_STATUS_CODES:
+            case AssertionNames.INVALID_HTTP_STATUS_CODES:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, InvalidHttpStatusCodesAssertion.class);
-            case SIMPLE_CONTAINS:
+            case AssertionNames.SIMPLE_CONTAINS:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, SimpleContainsAssertion.class);
-            case SIMPLE_NOT_CONTAINS:
+            case AssertionNames.SIMPLE_NOT_CONTAINS:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, SimpleNotContainsAssertion.class);
-            case X_PATH_MATCH:
+            case AssertionNames.X_PATH_MATCH:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, XPathContainsAssertion.class);
-            case X_QUERY_MATCH:
+            case AssertionNames.X_QUERY_MATCH:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, XQueryContainsAssertion.class);
-            case JSON_PATH_MATCH:
+            case AssertionNames.JSON_PATH_MATCH:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, JsonPathContentAssertion.class);
-            case JSON_PATH_COUNT:
+            case AssertionNames.JSON_PATH_COUNT:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, JsonPathCountAssertion.class);
-            case GROOVY_SCRIPT:
+            case AssertionNames.GROOVY_SCRIPT:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, GroovyScriptAssertion.class);
-            case RESPONSE_SLA:
+            case AssertionNames.RESPONSE_SLA:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, ResponseSLAAssertion.class);
-            case JDBC_STATUS:
+            case AssertionNames.JDBC_STATUS:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, JdbcStatusAssertion.class);
-            case JDBC_TIME_OUT:
+            case AssertionNames.JDBC_TIME_OUT:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, JdbcTimeoutAssertion.class);
-            case SCHEMA_COMPLIANCE:
+            case AssertionNames.SCHEMA_COMPLIANCE:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, SchemaComplianceAssertion.class);
-            case SOAP_FAULT:
+            case AssertionNames.SOAP_FAULT:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, SoapFaultAssertion.class);
-            case NOT_SOAP_FAULT:
+            case AssertionNames.NOT_SOAP_FAULT:
                 return TypeFactory.defaultInstance().constructSpecializedType(baseType, NotSoapFaultAssertion.class);
             default:
                 return null; //TypeFactory.defaultInstance().constructSpecializedType(baseType, PluginAssertion.class);
